+++ Modules/Appraisal/Appraisal/Application/Features/Appraisals/CreateLeaseAgreementLandAndBuildingProperty/CreateLeaseAgreementLandAndBuildingPropertyRequest.cs
@@ -131,7 +131,7 @@ public record CreateLeaseAgreementLandAndBuildingPropertyRequest(
     string? BuildingAreaUnit = null,
     decimal? UsableArea = null,
     // Building - Structure
-    int? NumberOfFloors = null,
+    decimal? NumberOfFloors = null,
     // Building - Style
     string? BuildingMaterialType = null,
     string? BuildingStyleType = null,
